+++ tests/Unit/States/PageStateTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Thinktomorrow\Chief\Tests\Unit\Resource\States;
+namespace Thinktomorrow\Chief\Tests\Unit\States;
 
+use Thinktomorrow\Chief\Tests\Unit\Forms\TestCase;
+use Thinktomorrow\Chief\Tests\Shared\Fakes\ArticlePage;
+use Thinktomorrow\Chief\ManagedModels\States\State\StateMachine;
 use Thinktomorrow\Chief\ManagedModels\States\PageState\PageState;
 use Thinktomorrow\Chief\ManagedModels\States\State\StateException;
-use Thinktomorrow\Chief\ManagedModels\States\State\StateMachine;
-use Thinktomorrow\Chief\Tests\Shared\Fakes\ArticlePage;
-use Thinktomorrow\Chief\Tests\Unit\Forms\TestCase;
 
 class PageStateTest extends TestCase
 {