+++ tests/Unit/States/StateMachine/StateMachineIntegrityTest.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine;
+namespace Thinktomorrow\Chief\Tests\Unit\States\StateMachine;
 
 use PHPUnit\Framework\TestCase;
-use Thinktomorrow\Chief\ManagedModels\States\State\StateException;
 use Thinktomorrow\Chief\ManagedModels\States\State\StateMachine;
-use Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine\Stubs\MalformedStateConfigStub;
-use Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine\Stubs\MissingStateConfigStub;
-use Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine\Stubs\OnlineStateStub;
-use Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine\Stubs\StatefulStub;
+use Thinktomorrow\Chief\ManagedModels\States\State\StateException;
+use Thinktomorrow\Chief\Tests\Unit\States\StateMachine\Stubs\StatefulStub;
+use Thinktomorrow\Chief\Tests\Unit\States\StateMachine\Stubs\OnlineStateStub;
+use Thinktomorrow\Chief\Tests\Unit\States\StateMachine\Stubs\MissingStateConfigStub;
+use Thinktomorrow\Chief\Tests\Unit\States\StateMachine\Stubs\MalformedStateConfigStub;
 
 class StateMachineIntegrityTest extends TestCase
 {