+++ tests/Unit/States/StateMachine/Stubs/OnlineStateStub.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Thinktomorrow\Chief\Tests\Unit\Resource\States\StateMachine\Stubs;
+namespace Thinktomorrow\Chief\Tests\Unit\States\StateMachine\Stubs;
 
 use Thinktomorrow\Chief\ManagedModels\States\State\State;